@@ -1,59 +1,64 @@
 class BooksController < ApplicationController
+  before_action :set_book, only: [:show, :edit, :update, :destroy]
+
   def index
     @books = Book.all
   end
 
-  def show
-    @book = Book.find(params[:id])
-    @comment = @book.comments.build(book_id: @book.id)
-    @comments = Comment.all
-    @author = @book.author(@book.author_id)
-  end
-  def edit
-    @book = Book.find(params[:id])
-  end
-
   def new
-    @books = Book.new
+    @book = Book.new
+    @authors = Author.all
   end
-  def update
-    @book = Book.find(params[:id])
 
-    @book.update(rating: params[:book][:rating])
+  def create
+    @book = Book.new(book_params)
 
     respond_to do |format|
-      format.html  # index.html.erb
-      format.turbo_stream  { render :json => @posts }
+      if @book.save
+        format.html { redirect_to @book, notice: 'Книга была успешно создана.' }
+        format.json { render :show, status: :created, location: @book }
+      else
+        format.html { render :new, status: :unprocessable_entity }
+        format.json { render json: @book.errors, status: :unprocessable_entity }
+      end
     end
   end
 
-  def characters
+  def show
+    @comments = @book.comments
+    @comment = Comment.new
   end
 
-  def staff
+  def edit
   end
 
-  def screenshots
+  def update
+    respond_to do |format|
+      if @book.update(book_params)
+        format.html { redirect_to @book, notice: 'Книга была успешно обновлена.' }
+        format.json { render :show, status: :ok, location: @book }
+      else
+        format.html { render :edit, status: :unprocessable_entity }
+        format.json { render json: @book.errors, status: :unprocessable_entity }
+      end
+    end
   end
 
-  def videos
-  end
-
-  def related
-  end
-
-  def chronology
-  end
-
-  def art
+  def destroy
+    @book.destroy
+    respond_to do |format|
+      format.html { redirect_to books_url, notice: 'Книга была успешно удалена.' }
+      format.json { head :no_content }
+    end
   end
 
   private
+
   def set_book
     @book = Book.find(params[:id])
   end
 
   def book_params
-    params.require(:book).permit(:title, :description, :author, :thumbnail, :user_id)
+    params.require(:book).permit(:title, :author_id, :description, :poster)
   end
 end
