@@ -13,6 +13,13 @@ class ApplicationController < ActionController::Base
 
   def configure_permitted_parameters
     devise_parameter_sanitizer.permit(:sign_up, keys: [:username, :email, :password])
-    devise_parameter_sanitizer.permit(:account_update, keys: [:username, :email, :password, :current_password])
+    devise_parameter_sanitizer.permit(:account_update, keys: [:username, :email, :password, :current_password,
+                                                              :first_name, :last_name])
+  end
+
+  def render_turbo_stream_with_flash(response = nil)
+    render turbo_stream: [response, turbo_stream.append('flash', partial: 'common/flash', locals: {
+      type: flash[:type], message: flash[:message]
+    })].compact
   end
 end
