@@ -1,68 +1,25 @@
 class CommentsController < ApplicationController
-  before_action :set_comment, only: %i[ show edit update destroy ]
+  before_action :set_book
 
-  # GET /comments or /comments.json
-  def index
-    @comments = Comment.all
-  end
-
-  # GET /comments/1 or /comments/1.json
-  def show
-  end
-
-  # GET /comments/new
-  def new
-    @comment = Comment.new
-  end
-
-  # GET /comments/1/edit
-  def edit
-  end
-
-  # POST /comments or /comments.json
   def create
-    @comment = Comment.new(comment_params)
+    @comment = @book.comments.build(comment_params)
+    @comment.user = current_user
 
-    respond_to do |format|
-      if @comment.save
-        format.html { redirect_to @comment.book, notice: "Comment was successfully created." }
-      else
-        format.html { render :new, status: :unprocessable_entity }
-      end
-    end
-  end
-
-  # PATCH/PUT /comments/1 or /comments/1.json
-  def update
-    respond_to do |format|
-      if @comment.update(comment_params)
-        format.html { redirect_to comment_url(@comment), notice: "Comment was successfully updated." }
-        format.json { render :show, status: :ok, location: @comment }
-      else
-        format.html { render :edit, status: :unprocessable_entity }
-        format.json { render json: @comment.errors, status: :unprocessable_entity }
-      end
-    end
-  end
-
-  # DELETE /comments/1 or /comments/1.json
-  def destroy
-    @comment.destroy
-
-    respond_to do |format|
-      format.html { redirect_to comments_url, notice: "Comment was successfully destroyed." }
-      format.json { head :no_content }
+    if @comment.save
+      redirect_to @book, notice: 'Комментарий был успешно добавлен.'
+    else
+      error_message = @comment.errors.full_messages.join(', ')
+      redirect_to @book, alert: "Ошибка при добавлении комментария: #{error_message}"
     end
   end
 
   private
-    # Use callbacks to share common setup or constraints between actions.
-    def set_comment
-      @comment = Comment.find(params[:id])
-    end
 
-    # Only allow a list of trusted parameters through.
+  def set_book
+    @book = Book.find(params[:book_id])
+  end
+
   def comment_params
-    params.require(:comment).permit(:book_id, :text)
+    params.require(:comment).permit(:text, :book_id)
   end
 end
