@@ -1,11 +1,14 @@
 import { Application } from "@hotwired/stimulus"
-import star_bar_controller from "./star_bar_controller";
+import TextareaAutogrow from 'stimulus-textarea-autogrow'
 
-const application = Application.start()
+// Start Stimulus
+const Stimulus = Application.start()
+Stimulus.register('textarea-autogrow', TextareaAutogrow)
 
-// Configure Stimulus development experience
-application.debug = false
-window.Stimulus   = application
+import Flash_controller from "./flash_controller.js";
+Stimulus.register("flash", Flash_controller);
 
-export { application }
+import Tiptap_controller from "./tiptap_controller.js";
+Stimulus.register("tiptap", Tiptap_controller);
 
+export { Stimulus }
