@@ -1,7 +1,8 @@
 class CreateComments < ActiveRecord::Migration[7.0]
   def change
     create_table :comments do |t|
-      t.integer :book_id
+      t.references :user, null: false, foreign_key: true
+      t.references :book, null: false, foreign_key: true
       t.text :text
 
       t.timestamps
