@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_02_24_114349) do
+ActiveRecord::Schema[7.1].define(version: 2024_07_17_130453) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
@@ -61,10 +61,22 @@ ActiveRecord::Schema[7.0].define(version: 2023_02_24_114349) do
   end
 
   create_table "comments", force: :cascade do |t|
-    t.integer "book_id"
+    t.bigint "user_id", null: false
+    t.bigint "book_id", null: false
     t.text "text"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.index ["book_id"], name: "index_comments_on_book_id"
+    t.index ["user_id"], name: "index_comments_on_user_id"
+  end
+
+  create_table "libraries", force: :cascade do |t|
+    t.bigint "user_id", null: false
+    t.bigint "book_id", null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["book_id"], name: "index_libraries_on_book_id"
+    t.index ["user_id"], name: "index_libraries_on_user_id"
   end
 
   create_table "users", force: :cascade do |t|
@@ -77,10 +89,16 @@ ActiveRecord::Schema[7.0].define(version: 2023_02_24_114349) do
     t.datetime "updated_at", null: false
     t.string "username"
     t.string "role"
+    t.string "first_name"
+    t.string "last_name"
     t.index ["email"], name: "index_users_on_email", unique: true
     t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
 
   add_foreign_key "active_storage_attachments", "active_storage_blobs", column: "blob_id"
   add_foreign_key "active_storage_variant_records", "active_storage_blobs", column: "blob_id"
+  add_foreign_key "comments", "books"
+  add_foreign_key "comments", "users"
+  add_foreign_key "libraries", "books"
+  add_foreign_key "libraries", "users"
 end
